requirements.txt
numpy==1.25.1
Pillow==10.0.0
